make ability to change title for selected film
