+++ app/src/main/java/com/vkochenkov/filmscatalog/activities/FilmsListActivity.kt
@@ -6,18 +6,19 @@ import android.view.View
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import com.vkochenkov.filmscatalog.data.DataStorage
 import com.vkochenkov.filmscatalog.R
-import com.vkochenkov.filmscatalog.for_recycler.FilmsAdapter
+import com.vkochenkov.filmscatalog.recycler.FilmsAdapter
 import com.vkochenkov.filmscatalog.model.Film
 
 class FilmsListActivity : AppCompatActivity(), View.OnClickListener {
 
     companion object {
         const val FILM = "FILM"
-        const val CHOSEN_FILM_TITLE = "CHOSEN_FILM_TITLE"
     }
 
-    private lateinit var filmsArr: Array<Film>
+    var filmsArr = DataStorage.filmsArr
+
     lateinit var filmsRecycler: RecyclerView
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -25,7 +26,6 @@ class FilmsListActivity : AppCompatActivity(), View.OnClickListener {
         setContentView(R.layout.activity_films_list)
 
         initFields()
-        initFilmsArray()
         initRecycler()
         setOnClickListeners()
     }
@@ -37,7 +37,7 @@ class FilmsListActivity : AppCompatActivity(), View.OnClickListener {
     private fun initRecycler() {
         filmsRecycler.layoutManager = LinearLayoutManager(this)
         filmsRecycler.adapter = FilmsAdapter(filmsArr) { film ->
-            openFilmActivity(film)
+            openSelectedFilmActivity(film)
         }
     }
 
@@ -47,56 +47,16 @@ class FilmsListActivity : AppCompatActivity(), View.OnClickListener {
         }
     }
 
-    //    override fun onSaveInstanceState(outState: Bundle) {
-//        super.onSaveInstanceState(outState)
-//        chosenFilmTitle?.let { outState.putInt(CHOSEN_FILM_TITLE, it) }
-//    }
-//
-//    override fun onRestoreInstanceState(savedInstanceState: Bundle) {
-//        super.onRestoreInstanceState(savedInstanceState)
-//        chosenFilmTitle = savedInstanceState.getInt(CHOSEN_FILM_TITLE)
-//        changeTitleColorForSelectedFilm()
-//    }
-//
     private fun setOnClickListeners() {
         findViewById<View>(R.id.btn_share).setOnClickListener(this)
     }
 
-    private fun initFilmsArray() {
-        filmsArr = arrayOf(
-            Film(R.string.film_title_1, R.string.film_description_1, R.drawable.im_film_1),
-            Film(R.string.film_title_2, R.string.film_description_2, R.drawable.im_film_2),
-            Film(R.string.film_title_3, R.string.film_description_3, R.drawable.im_film_3),
-            Film(R.string.film_title_4, R.string.film_description_4, R.drawable.im_film_4),
-            Film(R.string.film_title_5, R.string.film_description_5, R.drawable.im_film_5),
-            Film(R.string.film_title_6, R.string.film_description_6, R.drawable.im_film_6)
-        )
-    }
-
-//    private fun clickOnFilmDetails(film: Film, titleId: Int) {
-//        chosenFilmTitle = titleId
-//        changeTitleColorForSelectedFilm()
-//        openFilmActivity(film)
-//    }
-
-    private fun openFilmActivity(film: Film) {
+    private fun openSelectedFilmActivity(film: Film) {
         val intent = Intent(this, FilmInfoActivity::class.java).apply {
             putExtra(FILM, film)
         }
         startActivity(intent)
     }
-
-//    private fun changeTitleColorForSelectedFilm() {
-//        val accentColor = ContextCompat.getColor(this, R.color.colorAccent)
-//        val defaultColor = ContextCompat.getColor(this, R.color.colorBlack)
-//        for (element in titlesArr) {
-//            if (element == chosenFilmTitle) {
-//                findViewById<TextView>(element).setTextColor(accentColor)
-//            } else {
-//                findViewById<TextView>(element).setTextColor(defaultColor)
-//            }
-//        }
-//    }
 
     private fun shareFriends() {
         val intent = Intent(Intent.ACTION_SEND).apply {
