+++ app/src/main/java/com/vkochenkov/filmscatalog/recycler/FilmViewHolder.kt
@@ -1,9 +1,10 @@
-package com.vkochenkov.filmscatalog.for_recycler
+package com.vkochenkov.filmscatalog.recycler
 
 import android.view.View
 import android.widget.Button
 import android.widget.ImageView
 import android.widget.TextView
+import androidx.core.content.ContextCompat
 import androidx.recyclerview.widget.RecyclerView
 import com.vkochenkov.filmscatalog.R
 import com.vkochenkov.filmscatalog.model.Film
@@ -16,5 +17,10 @@ class FilmViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
     fun bind(item: Film) {
         filmTitle.text = itemView.context.getString(item.titleRes)
         filmImage.setImageResource(item.imageRes)
+        if (item.selected) {
+            filmTitle.setTextColor(ContextCompat.getColor(itemView.context, R.color.colorAccent))
+        } else {
+            filmTitle.setTextColor(ContextCompat.getColor(itemView.context, R.color.colorBlack))
+        }
     }
 }
